pyproject.toml
[tool.poetry]
name = "tema-sport"
version = "0.0.1"
description = ""
authors = ["trifonovtema <artyom85@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
aiokafka = "^0.11.0"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"
redis = "^5.0.8"
asyncpg = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pre-commit = "^3.8.0"
black = "^24.8.0"
poetry = "^1.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
